pg/pyproject.toml
[tool.poetry]
name = "pangea"
version = "0.1.0"
description = ""
authors = ["Damian Krystkiewicz <damian.krystkiewicz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.2"
django-environ = "^0.4.5"
tqdm = "^4.45.0"
pexpect = "^4.8.0"
loguru = "^0.4.1"
fire = "^0.3.1"
flake8 = "^3.7.9"
jinja2 = "^2.11.2"
black = "^19.10b0"
pytest-socket = "^0.3.4"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
mypy = "^0.770"
pep8-naming = "^0.10.0"
pytest = "^5.4.1"
envo = {path = "../envo"}
pytest-mock = "^3.1.0"
loguru_caplog = "^0.1.3"

[tool.poetry.scripts]
pangea = "pangea.scripts:_main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
